+++ cpskin/workflow/tests/test_profiles.py
@@ -20,6 +20,7 @@ class TestProfiles(unittest.TestCase):
         workflow = getToolByName(portal, 'portal_workflow')
         self.assertTrue('cpskin_workflow' in workflow)
         self.assertTrue('cpskin_moderation_workflow' in workflow)
+        self.assertTrue('cpskin_collective_contact_workflow' in workflow)
         self.assertFalse('cpskin_readonly_workflow' in workflow)
 
     def test_members_profile(self):
@@ -34,6 +35,7 @@ class TestProfiles(unittest.TestCase):
         workflow = getToolByName(portal, 'portal_workflow')
         self.assertFalse('cpskin_workflow' in workflow)
         self.assertFalse('cpskin_moderation_workflow' in workflow)
+        self.assertFalse('cpskin_collective_contact_workflow' in workflow)
 
     def test_complete_uninstall(self):
         portal = self.layer['portal']
@@ -41,10 +43,12 @@ class TestProfiles(unittest.TestCase):
         applyProfile(portal, 'cpskin.workflow:members-configuration')
         self.assertTrue('cpskin_workflow' in workflow)
         self.assertTrue('cpskin_moderation_workflow' in workflow)
+        self.assertTrue('cpskin_collective_contact_workflow' in workflow)
         self.assertTrue('cpskin_readonly_workflow' in workflow)
         applyProfile(portal, 'cpskin.workflow:uninstall')
         self.assertFalse('cpskin_workflow' in workflow)
         self.assertFalse('cpskin_moderation_workflow' in workflow)
+        self.assertFalse('cpskin_collective_contact_workflow' in workflow)
         self.assertFalse('cpskin_readonly_workflow' in workflow)
 
     def test_uninstall_workflows_with_public_object(self):